projects/screenshot capture/requirements.txt
Pillow
schedule